+++ bokeh/colors/color.py
@@ -23,12 +23,11 @@ log = logging.getLogger(__name__)
 # Standard library imports
 import colorsys
 from abc import ABCMeta, abstractmethod
-from typing import TYPE_CHECKING, Type, TypeVar
+from math import sqrt
+from typing import Type, TypeVar
 
-## Bokeh imports
-if TYPE_CHECKING:
-    from .hsl import HSL
-    from .rgb import RGB
+# Bokeh imports
+from ..util.deprecation import deprecated
 
 #-----------------------------------------------------------------------------
 # Globals and constants
@@ -125,7 +124,7 @@ class Color(metaclass=ABCMeta):
         *Subclasses must implement this method.*
 
         Args:
-            value (:class:`~bokeh.colors.rgb.RGB`) :
+            value (:class:`~bokeh.colors.RGB`) :
                 A color to convert from RGB
 
         Returns:
@@ -187,10 +186,259 @@ class Color(metaclass=ABCMeta):
         *Subclasses must implement this method.*
 
         Returns:
-            :class:`~bokeh.colors.rgb.RGB`
+            :class:`~bokeh.colors.RGB`
 
         '''
         raise NotImplementedError
+
+class RGB(Color):
+    ''' Represent colors by specifying their Red, Green, and Blue channels.
+
+    Alpha values may also optionally be provided. Otherwise, alpha values
+    default to 1.
+
+    '''
+
+    def __init__(self, r: int, g: int, b: int, a: float = 1.0) -> None:
+        '''
+
+        Args:
+            r (int) :
+                The value for the red channel in [0, 255]
+
+            g (int) :
+                The value for the green channel in [0, 255]
+
+            b (int) :
+                The value for the blue channel in [0, 255]
+
+            a (float, optional) :
+                An alpha value for this color in [0, 1] (default: 1.0)
+
+        '''
+        self.r = r
+        self.g = g
+        self.b = b
+        self.a = a
+
+    def copy(self) -> RGB:
+        ''' Return a copy of this color value.
+
+        Returns:
+            :class:`~bokeh.colors.RGB`
+
+        '''
+        return RGB(self.r, self.g, self.b, self.a)
+
+    @classmethod
+    def from_hsl(cls, value: HSL) -> RGB:
+        ''' Create an RGB color from an HSL color value.
+
+        Args:
+            value (HSL) :
+                The HSL color to convert.
+
+        Returns:
+            :class:`~bokeh.colors.RGB`
+
+        '''
+        return value.to_rgb()
+
+    @classmethod
+    def from_rgb(cls, value: RGB) -> RGB:
+        ''' Copy an RGB color from another RGB color value.
+
+        Args:
+            value (:class:`~bokeh.colors.RGB`) :
+                The RGB color to copy.
+
+        Returns:
+            :class:`~bokeh.colors.RGB`
+
+        '''
+        return value.copy()
+
+    def to_css(self) -> str:
+        ''' Generate the CSS representation of this RGB color.
+
+        Returns:
+            str, ``"rgb(...)"`` or ``"rgba(...)"``
+
+        '''
+        if self.a == 1.0:
+            return f"rgb({self.r}, {self.g}, {self.b})"
+        else:
+            return f"rgba({self.r}, {self.g}, {self.b}, {self.a})"
+
+    def to_hex(self) -> str:
+        ''' Return a hex color string for this RGB color.
+
+        Any alpha value on this color is discarded, only hex color strings for
+        the RGB components are returned.
+
+        Returns:
+            str, ``"#RRGGBB"``
+
+        '''
+        return "#%02X%02X%02X" % (self.r, self.g, self.b)
+
+    def to_hsl(self) -> HSL:
+        ''' Return a corresponding HSL color for this RGB color.
+
+        Returns:
+            :class:`~bokeh.colors.HSL`
+
+        '''
+        h, l, s = colorsys.rgb_to_hls(float(self.r)/255, float(self.g)/255, float(self.b)/255)
+        return HSL(round(h*360), s, l, self.a)
+
+    def to_rgb(self) -> RGB:
+        ''' Return a RGB copy for this RGB color.
+
+        Returns:
+            :class:`~bokeh.colors.RGB`
+
+        '''
+        return self.copy()
+
+    @property
+    def brightness(self) -> float:
+        """ Perceived brightness of a color in [0, 1] range. """
+        # http://alienryderflex.com/hsp.html
+        r, g, b = self.r, self.g, self.b
+        return sqrt(0.299*r**2 + 0.587*g**2 + 0.114*b**2)/255
+
+class HSL(Color):
+    ''' Represent colors by specifying their Hue, Saturation, and lightness.
+
+    Alpha values may also optionally be provided. Otherwise, alpha values
+    default to 1.
+
+    .. warning::
+        HSL is deprecated as of Bokeh 2.3.1 and will be removed in a future
+        release. Use RGB or named colors instead.
+
+    '''
+
+    def __init__(self, h: float, s: float, l: float, a: float = 1.0) -> None:
+        '''
+
+        Args:
+            h (int) :
+                The Hue, in [0, 360]
+
+            s (int) :
+                The Saturation, in [0, 1]
+
+            l (int) :
+                The lightness, in [0, 1]
+
+            a (float, optional) :
+                An alpha value for this color in [0, 1] (default: 1.0)
+
+        '''
+        deprecated((2, 3, 1), "HSL()", "RGB() or named colors")
+        self.h = h
+        self.s = s
+        self.l = l
+        self.a = a
+
+    def copy(self) -> HSL:
+        ''' Return a copy of this color value.
+
+        Returns:
+            :class:`~bokeh.colors.HSL`
+
+        '''
+        return HSL(self.h, self.s, self.l, self.a)
+
+    def darken(self, amount: float) -> HSL:
+        ''' Darken (reduce the luminance) of this color.
+
+        Args:
+            amount (float) :
+                Amount to reduce the luminance by (clamped above zero)
+
+        Returns:
+            :class:`~bokeh.colors.HSL`
+
+        '''
+        return self.lighten(-amount)
+
+    @classmethod
+    def from_hsl(cls, value: HSL) -> HSL:
+        ''' Copy an HSL color from another HSL color value.
+
+        Args:
+            value (HSL) :
+                The HSL color to copy.
+
+        Returns:
+            :class:`~bokeh.colors.hsl.HSL`
+
+        '''
+        return value.copy()
+
+    @classmethod
+    def from_rgb(cls, value: RGB) -> HSL:
+        ''' Create an HSL color from an RGB color value.
+
+        Args:
+            value (:class:`~bokeh.colors.RGB`) :
+                The RGB color to convert.
+
+        Returns:
+            :class:`~bokeh.colors.HSL`
+
+        '''
+        return value.to_hsl()
+
+    def to_css(self) -> str:
+        ''' Generate the CSS representation of this HSL color.
+
+        Returns:
+            str, ``"hsl(...)"`` or ``"hsla(...)"``
+
+        '''
+        if self.a == 1.0:
+            return f"hsl({self.h}, {self.s*100}%, {self.l*100}%)"
+        else:
+            return f"hsla({self.h}, {self.s*100}%, {self.l*100}%, {self.a})"
+
+    def to_hsl(self) -> HSL:
+        ''' Return a HSL copy for this HSL color.
+
+        Returns:
+            :class:`~bokeh.colors.HSL`
+
+        '''
+        return self.copy()
+
+    def to_rgb(self) -> RGB:
+        ''' Return a corresponding :class:`~bokeh.colors.RGB` color for
+        this HSL color.
+
+        Returns:
+            :class:`~bokeh.colors.RGB`
+
+        '''
+        r, g, b = colorsys.hls_to_rgb(float(self.h)/360, self.l, self.s)
+        return RGB(round(r*255), round(g*255), round(b*255), self.a)
+
+    def lighten(self, amount: float) -> HSL:
+        ''' Lighten (increase the luminance) of this color.
+
+        Args:
+            amount (float) :
+                Amount to increase the luminance by (clamped above zero)
+
+        Returns:
+            :class:`~bokeh.colors.HSL`
+
+        '''
+        hsl = self.copy()
+        hsl.l = self.clamp(hsl.l + amount, 1)
+        return self.from_hsl(hsl)
 
 #-----------------------------------------------------------------------------
 # Dev API
